Auction testing

+++ types/State/Market/AuctionState.cs
@@ -6,10 +6,10 @@ namespace Echo.State.Market
 	{
 		public ulong ObjectId { get; set; }
 		public string Name { get; set; }
-		public int BlockSize { get; set; }
+		public uint BlockSize { get; set; }
 		public long Expires { get; set; }
 		public int PricePerUnit { get; set; }
-		public int Range { get; set; }
+		public double Range { get; set; }
 
 		public ObjectReference Owner { get; set; }
 		public ObjectReference Trader { get; set; }
